renaming hooks; fixing styles after importing ant design
