+++ client/package.json
@@ -14,6 +14,7 @@
     "@types/react-dom": "^17.0.2",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.25",
+    "antd": "^4.21.7",
     "axios": "^0.27.2",
     "exif-js": "^2.3.0",
     "lodash": "^4.17.21",
