+++ client/src/components/LoadingSpinner/spinner.tsx
@@ -1,0 +1,3 @@
+import {Spinner} from './spinner.styles';
+
+export default Spinner;
