+++ client/src/pages/Album/album.styles.ts
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 export const Wrapper = styled.div`
     max-width: var(--maxWidthDesktop);  //control the width of the grid only displaying 5 in a row, maybe play with this if I decide to go full-screen with it
     margin: 0 auto;
-    padding: 0 20px 20px 20px;
+    padding: 20px 20px 20px 20px;
 
     h3{
         color: var(--medGrey);
