+++ client/src/pages/Photos/photos.styles.ts
@@ -3,117 +3,43 @@ import styled from 'styled-components';
 export const Wrapper = styled.div`
     max-width: var(--maxWidthDesktop);  //control the width of the grid only displaying 5 in a row, maybe play with this if I decide to go full-screen with it
     margin: 0 auto;
-    padding: 50px 20px 20px 20px;
-    width: 50%;
-    min-width: 800px; 
+    padding: 20px 20px 20px 20px;
 
-    .button-container{
-        display: flex;
-        justify-content: flex-end;
-        margin-top: 20px;
+    h3{
+        color: var(--medGrey);
+        font-weight: 500;
+        @media screen and (max-width: 768px){
+            font-size: var(--fontBig);
+        }
     }
 
-    button {
-        display:inline-block;
-        padding: 12px;
-        color: white;
-        /* background-color: var(--red); */
-        border: none;
-        font-size: medium;
-        border-radius: 5px;
-        cursor: pointer;
-        text-transform: uppercase;
-        letter-spacing: 1px;
-    }
-
-    .cancel-button {
-        background-color: var(--smoke);
-        color: var(--snow);
-        margin-right: 20px;
-    }
-
-    .cancel-button:hover {
-        background-color: var(--slate);
-    }
-
-    .accept-button {
-        background-color: var(--darkRed);
-        color: var(--snow);
-    }
-
-    .accept-button:hover {
-        background-color: var(--red);
-    }
-
-    .thumb-wrapper{
-        padding: 10px 0 5px 10px;
-        min-height: 40px;
-        border: 1px solid var(--fog);
+    .divider {
+        border-bottom: 1px solid var(--steam);
+        margin-bottom: 20px;
     }
 `;
+
 
 export const Content = styled.div`
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 20px;
-    border-width: 2px;
-    border-radius: 2px;
-    border-color: var(--fog);
-    border-style: dashed;
-    background-color: #fafafa;
-    color: #bdbdbd;
-    outline: none;
-    transition: border .24s ease-in-out;
+    display: grid;
+    // set to repeat the columns, and then set it to auto fill, and then when its 200px wide it cant' go lower so it removes 1 column instead to make it responsive for mobile
+    grid-template-columns: repeat(auto-fill, minmax(400px, 1fr)); //css grid syntax here for a responsive grid
+    grid-gap: 1.3rem; //spacing between rows/columns
 
-    h2 {
-        margin-top: 66px;
+    /* @media screen and (max-width: 768px) {
+        grid-template-columns: repeat(auto-fill, minmax(150px, 1fr)); //css grid syntax here for a responsive grid
     }
-`;
 
-export const UploadImage = styled.img`
-    width: 150px;
-    padding: 40px 0 60px 0;
+    @media screen and (max-width: 500px) {
+        grid-template-columns: repeat(auto-fill, minmax(100px, 1fr)); //css grid syntax here for a responsive grid
+    } */
+
+    .fake-content{
+        width: 300px;
+        height: 300px;
+        border-radius: 5px;
+        background-color: var(--steam);
+    }
 
 `;
-
-
-export const ThumbsContainer = styled.aside`
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    //margin-top: 16px;
-
-    h4{
-        margin: 0;
-        margin-top: 10px;
-    }
-    .thumb {
-        display: inline-flex;
-        border-radius: 2px;
-        border: 1px solid #eaeaea;
-        margin-bottom: 8px;
-        margin-right: 8px;
-        width: 100px;
-        height: 100px;
-        padding: 4px;
-        box-sizing: border-box;
-    }
-
-    .thumb-inner {
-        display: flex;
-        min-width: 0;
-        overflow: hidden;
-        justify-content: center;
-    }
-
-    .thumb-image {
-        display: block;
-        width: auto;
-        height: 100%;
-    }
-`;
-
-
