--- client/src/pages/Upload/upload.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import { Wrapper, Content } from './upload.styles';
-
-const Upload = () => {
-
-    return (
-        <></>
-    )
-
-}
-
-export default Upload;
